+++ src/components/datagrid/body/data_grid_cell.test.tsx
@@ -180,12 +180,17 @@ describe('EuiDataGridCell', () => {
   });
 
   describe('componentDidUpdate', () => {
-    it('resets cell props when the cell columnId changes', () => {
+    it('resets cell props when the cell is moved (columnId) or sorted (rowIndex)', () => {
       const setState = jest.spyOn(EuiDataGridCell.prototype, 'setState');
       const component = mount(<EuiDataGridCell {...requiredProps} />);
 
       component.setProps({ columnId: 'newColumnId' });
       expect(setState).toHaveBeenCalledWith({ cellProps: {} });
+      expect(setState).toHaveBeenCalledTimes(1);
+
+      component.setProps({ rowIndex: 1 });
+      expect(setState).toHaveBeenCalledWith({ cellProps: {} });
+      expect(setState).toHaveBeenCalledTimes(2);
     });
 
     it("handles the cell popover by forwarding the cell's DOM node and contents to the parent popover context", () => {