+++ src/components/datagrid/body/data_grid_cell.tsx
@@ -301,7 +301,10 @@ export class EuiDataGridCell extends Component<
       this.handleCellPopover();
     }
 
-    if (this.props.columnId !== prevProps.columnId) {
+    if (
+      this.props.columnId !== prevProps.columnId ||
+      this.props.rowIndex !== prevProps.rowIndex
+    ) {
       this.setCellProps({});
     }
   }